Fix modules dependencies.
