+++ modules/schema/build.gradle
@@ -29,6 +29,7 @@ dependencies {
     api project(':ignite-binary-tuple')
     implementation project(':ignite-api')
     implementation project(':ignite-bytecode')
+    implementation project(':ignite-catalog')
     implementation project(':ignite-core')
     implementation project(':ignite-configuration-api')
     implementation project(':ignite-metastorage-api')
