+++ modules/schema/src/main/java/org/apache/ignite/internal/schema/CatalogDescriptorUtils.java
@@ -15,18 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.catalog.descriptors;
+package org.apache.ignite.internal.schema;
 
 import static java.util.stream.Collectors.toList;
 import static org.apache.ignite.internal.catalog.descriptors.CatalogColumnCollation.ASC_NULLS_LAST;
 import static org.apache.ignite.internal.catalog.descriptors.CatalogColumnCollation.DESC_NULLS_FIRST;
 
-import java.lang.reflect.Method;
 import java.util.List;
 import org.apache.ignite.internal.catalog.commands.DefaultValue;
-import org.apache.ignite.internal.distributionzones.configuration.DistributionZoneView;
-import org.apache.ignite.internal.schema.NativeType;
-import org.apache.ignite.internal.schema.NativeTypes;
+import org.apache.ignite.internal.catalog.descriptors.CatalogColumnCollation;
+import org.apache.ignite.internal.catalog.descriptors.CatalogHashIndexDescriptor;
+import org.apache.ignite.internal.catalog.descriptors.CatalogIndexColumnDescriptor;
+import org.apache.ignite.internal.catalog.descriptors.CatalogIndexDescriptor;
+import org.apache.ignite.internal.catalog.descriptors.CatalogSortedIndexDescriptor;
+import org.apache.ignite.internal.catalog.descriptors.CatalogTableColumnDescriptor;
+import org.apache.ignite.internal.catalog.descriptors.CatalogTableDescriptor;
 import org.apache.ignite.internal.schema.configuration.ColumnTypeView;
 import org.apache.ignite.internal.schema.configuration.ColumnView;
 import org.apache.ignite.internal.schema.configuration.ConfigurationToSchemaDescriptorConverter;
@@ -42,13 +45,13 @@ import org.apache.ignite.internal.schema.configuration.index.IndexColumnView;
 import org.apache.ignite.internal.schema.configuration.index.SortedIndexView;
 import org.apache.ignite.internal.schema.configuration.index.TableIndexConfigurationSchema;
 import org.apache.ignite.internal.schema.configuration.index.TableIndexView;
-import org.apache.ignite.internal.schema.configuration.storage.DataStorageView;
 
 /**
  * Helper class for working with catalog descriptors.
  */
 // TODO: IGNITE-19499 Get rid of the table configuration
 // TODO: IGNITE-19500 Get rid of the index configuration
+@Deprecated(forRemoval = true)
 public class CatalogDescriptorUtils {
     /**
      * Converts a table configuration to a catalog table descriptor.
@@ -152,25 +155,6 @@ public class CatalogDescriptorUtils {
         }
     }
 
-    /**
-     * Converts a distribution zone configuration to a Distribution zone descriptor.
-     *
-     * @param config Distribution zone configuration.
-     */
-    public static CatalogZoneDescriptor toZoneDescriptor(DistributionZoneView config) {
-        return new CatalogZoneDescriptor(
-                config.zoneId(),
-                config.name(),
-                config.partitions(),
-                config.replicas(),
-                config.dataNodesAutoAdjust(),
-                config.dataNodesAutoAdjustScaleUp(),
-                config.dataNodesAutoAdjustScaleDown(),
-                config.filter(),
-                toDataStorageDescriptor(config.dataStorage())
-        );
-    }
-
     private static CatalogTableColumnDescriptor toTableColumnDescriptor(ColumnView config) {
         ColumnTypeView typeConfig = config.type();
 
@@ -210,22 +194,5 @@ public class CatalogDescriptorUtils {
         CatalogColumnCollation collation = config.asc() ? ASC_NULLS_LAST : DESC_NULLS_FIRST;
 
         return new CatalogIndexColumnDescriptor(config.name(), collation);
-    }
-
-    // TODO: IGNITE-19719 Fix it
-    private static CatalogDataStorageDescriptor toDataStorageDescriptor(DataStorageView config) {
-        String dataRegion;
-
-        try {
-            Method dataRegionMethod = config.getClass().getMethod("dataRegion");
-
-            dataRegionMethod.setAccessible(true);
-
-            dataRegion = (String) dataRegionMethod.invoke(config);
-        } catch (ReflectiveOperationException e) {
-            dataRegion = e.getMessage();
-        }
-
-        return new CatalogDataStorageDescriptor(config.name(), dataRegion);
     }
 }
