+++ modules/storage-api/src/main/java/org/apache/ignite/internal/storage/index/StorageSortedIndexDescriptor.java
@@ -18,7 +18,7 @@
 package org.apache.ignite.internal.storage.index;
 
 import static java.util.stream.Collectors.toList;
-import static org.apache.ignite.internal.catalog.descriptors.CatalogDescriptorUtils.getNativeType;
+import static org.apache.ignite.internal.schema.CatalogDescriptorUtils.getNativeType;
 
 import java.util.List;
 import org.apache.ignite.internal.catalog.descriptors.CatalogColumnCollation;
